Add interactive key rebind tool to SDL2 frontend
This involved writing a slightly messy function that updates an .ini file
while keeping its comments and whitespace intact. It seems to do a good job.

This tool isn't *great*, right now -- but issue #2313 prevents me from
rendering any text in the SDL window that pops up to prompt the user for a
single keypress.